+++ capacitor.config.ts
@@ -1,18 +1,21 @@
 import type { CapacitorConfig } from '@capacitor/cli';
 
 const config: CapacitorConfig = {
-  appId: 'com.empresa.baterponto',
-  appName: 'Bater Ponto',
+  appId: 'com.empresa.pontofácil',
+  appName: 'Ponto Fácil',
   webDir: 'www',
   server: {
     androidScheme: 'https'
   },
   plugins: {
     Camera: {
-      permissions: ['camera', 'photos']
+      permissions: ['camera', 'photos', 'storage']
     },
     Geolocation: {
       permissions: ['location']
+    },
+    Filesystem: {
+      permissions: ['storage']
     }
   }
 };